+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Add globally usable css class `polar-tooltip` intended to update the appearance of `v-tooltip`s to be more visible.
+
 ## 1.1.0
 
 - Feature: Add core state variable for map's center position.